Extracted strategies classes to separate files
